+++ src/whispr_server.cpp
@@ -1,4 +1,4 @@
-#include "server_socket.h"
+#include "../include/server_socket.h"
 
 int main() {
   Socket();
